requirements.txt
appnope==0.1.0
asgiref==3.2.10
attrs==19.3.0
autobahn==20.6.2
Automat==20.2.0
backcall==0.2.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
decorator==4.4.2
dj-database-url==0.5.0
Django==3.0.8
django-admin-autocomplete-filter==0.5
django-heroku==0.3.1
factory-boy==2.12.0
Faker==4.1.1
flake8==3.8.3
flake8-isort==3.0.1
hyperlink==19.0.0
idna==2.10
incremental==17.5.0
ipython==7.16.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.1
mccabe==0.6.1
more-itertools==8.4.0
packaging==20.4
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.5
psycopg2==2.8.5
ptyprocess==0.6.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.6.1
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytest==5.4.3
pytest-django==3.9.0
python-dateutil==2.8.1
python-env==1.0.0
pytz==2020.1
requests==2.24.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
testfixtures==6.14.1
text-unidecode==1.3
toml==0.10.1
traitlets==4.3.3
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.9
wcwidth==0.2.5
whitenoise==5.1.0
zope.interface==5.1.0
